requirements.txt
colorama==0.4.6
greenlet==3.0.3
joblib==1.4.2
nats-py==2.7.2
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.0
scipy==1.13.0
six==1.16.0
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
redis
